[GSoC 2026] M2.14 - Fix #22539: Prevent duplicate functions in playwright acceptance tests (#27058)
* Added linter for duplicate function

* Added runtime check

* Fixed duplicate functions

* Correcting the runtime check

* Extracted curriculum-admin to topic-manager

* Removed one function from navigation-utils.ts
