+++ core/tests/playwright-acceptance-tests/specs/community-library-browser/subscribe-to-a-favourite-creator.spec.ts
@@ -83,7 +83,7 @@ test.describe('Community Library Browser', function () {
       'Story 1',
       'Chapter 1'
     );
-    await communityLibraryBrowser.continueToNextCard();
+    await communityLibraryBrowser.continueToNextCardAsLoggedOutUser();
 
     // Subscribe to creator.
     await communityLibraryBrowser.openLessonInfoModal();
