+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/choose-new-lesson-to-play-from-learner-dashboard.spec.ts
@@ -141,7 +141,7 @@ test.describe('Logged-In Learner', function () {
     );
     await loggedInLearner.page.waitForLoadState('networkidle');
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     // TODO(#20869): A flaky behaviour is observed due to issue in the backend.
     // Even after completing the lesson, the node isn't marked as completed.
     // Once fixed, uncomment the below code.
@@ -154,7 +154,7 @@ test.describe('Logged-In Learner', function () {
     // await loggedInLearner.resumeLessonFromLearnerDashboard(
     //   'Chapter 2: Test Chapter 2'
     // );
-    // await loggedInLearner.continueToNextCard();
+    // await loggedInLearner.continueToNextCardAsLoggedOutUser();
     // await loggedInLearner.navigateToLearnerDashboard();
 
     // await loggedInLearner.expectLearnSomethingNewInLDToBeEmpty()
