+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/manage-classroom-progress-in-home-learner-dashboard.spec.ts
@@ -126,7 +126,7 @@ test.describe('Logged-In Learner', function () {
   });
 
   test('should have the correct tab title, available sections on landing and Sidebar should contain these items in this order from top to bottom: Profile picture, "Home" button, "Goals" button, "Progress" button', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectSidebarTabToBeActiveAndContainButtonsInOrder(
       'Home'
     );
@@ -157,29 +157,33 @@ test.describe('Logged-In Learner', function () {
 
   test('should navigate directly to math classroom', async function () {
     await loggedInLearner.navigateToClassroomFromLearnerDashboard('math');
-    await loggedInLearner.expectToBeOnPage('learn/math');
+    await loggedInLearner.expectToBeOnPageAsLoggedInUser('learn/math');
     await loggedInLearner.expectScreenshotToMatch('mathClassroomPage');
     showMessage('Navigated to math classroom from learner dashboard.');
   });
 
   test('should navigate directly to the Place Values topic in the math classroom', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToTopicPageByCard('Place Values');
-    await loggedInLearner.expectToBeOnPage('learn/math/place-values');
+    await loggedInLearner.expectToBeOnPageAsLoggedInUser(
+      'learn/math/place-values'
+    );
     await loggedInLearner.expectScreenshotToMatch('placeValuesTopicPage');
     showMessage('Navigated to Place Values topic from learner dashboard.');
   });
 
   test('should display in-progress and recommended lessons after starting a lesson', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToTopicPageByCard('Place Values');
-    await loggedInLearner.expectToBeOnPage('learn/math/place-values');
+    await loggedInLearner.expectToBeOnPageAsLoggedInUser(
+      'learn/math/place-values'
+    );
     await loggedInLearner.selectChapterWithinStoryToLearn(
       "Jamie's Adventures in the Arcade",
       'What are the Place Values'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     // Did not finish the chapter,So still in  In-progress section.
     await loggedInLearner.expectElementsToBePresentInRLD(
       ['Continue where you left off', 'Learn Something New'],
@@ -214,18 +218,18 @@ test.describe('Logged-In Learner', function () {
 
   test('should not recommend any lessons if currently on last lesson', async function () {
     test.setTimeout(480000); // Takes longer than default timeout.
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToLessonByCard(
       'Lessons in progress',
       'Chapter 1: What are the Place Values'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectScreenshotToMatch(
       'learnerDashboardHomeTabWithLessonsInProgresschapter2AndRecommendedForYouChapter3'
     );
@@ -244,13 +248,13 @@ test.describe('Logged-In Learner', function () {
       'Lessons in progress',
       'Chapter 2: Find the Value of a Number'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectScreenshotToMatch(
       'learnerDashboardHomeTabWithLessonsInProgresschapter3AndRecommendedForYouChapter4'
     );
@@ -270,13 +274,13 @@ test.describe('Logged-In Learner', function () {
       'Lessons in progress',
       'Chapter 3: Comparing Numbers'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectScreenshotToMatch(
       'learnerDashboardHomeTabWithLessonsInProgresschapter4AndRecommendedForYouChapter5'
     );
@@ -296,12 +300,12 @@ test.describe('Logged-In Learner', function () {
       'Lessons in progress',
       'Chapter 4: Rounding Numbers part 1'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectScreenshotToMatch(
       'learnerDashboardHomeTabWithLessonsInProgresschapter5AndRecommendedForYouChapter6'
     );
@@ -315,13 +319,13 @@ test.describe('Logged-In Learner', function () {
       'Lessons in progress',
       'Chapter 5: Rounding Numbers part 2'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectScreenshotToMatch(
       'learnerDashboardHomeTabWithLessonsInProgresschapter6AndNoRecommendedForYouChapter'
     );
