+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/manage-in-progress-and-completed-lessons-and-skill.spec.ts
@@ -116,7 +116,7 @@ test.describe('Logged-in Learner', function () {
   });
 
   test('should display empty progress message when no lessons are in progress', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.expectSidebarTabToBeActiveAndContainButtonsInOrder(
       'Home'
     );
@@ -129,13 +129,13 @@ test.describe('Logged-in Learner', function () {
   });
 
   test('should select "Or Explore All Lessons in Classroom" button and navigate to /learn/math', async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectClassroomButtonOnRedesignedLearnerDashboardToBePresent(
       true
     );
     await loggedInLearner.navigateThroughClassroomButtonOnRLD();
-    await loggedInLearner.expectToBeOnPage('/learn/math');
+    await loggedInLearner.expectToBeOnPageAsLoggedInUser('/learn/math');
   });
 
   test('should select Place Values Topic and play "Chapter 1: What are the Place Values?" but do not finish and see It in Progress Section', async function () {
@@ -144,9 +144,9 @@ test.describe('Logged-in Learner', function () {
       "Jamie's Adventures in the Arcade",
       'What are the Place Values'
     );
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectScreenshotToMatch(
@@ -171,13 +171,13 @@ test.describe('Logged-in Learner', function () {
       'Classroom Lessons',
       'Chapter 1: What are the Place Values'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectScreenshotToMatch(
       'ProgressSectionInProgressWithOnlyChapter02'
@@ -185,7 +185,7 @@ test.describe('Logged-in Learner', function () {
   });
 
   test("should complete all the lessons of Place Value's Story and see Chapter 1 in the Completed Lessons section", async function () {
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectLessonCardProgressToBe(
@@ -199,13 +199,13 @@ test.describe('Logged-in Learner', function () {
       'Classroom Lessons',
       'Chapter 2: Find the Value of a Number'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
     await loggedInLearner.expectLessonCardProgressToBe(
       'Classroom Lessons',
@@ -218,13 +218,13 @@ test.describe('Logged-in Learner', function () {
       'Classroom Lessons',
       'Chapter 3: Comparing Numbers'
     );
-    await loggedInLearner.continueToNextCard();
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
     await loggedInLearner.expectExplorationCompletionToastMessage(
       'Congratulations for completing this lesson!'
     );
 
-    await loggedInLearner.navigateToLearnerDashboard();
+    await loggedInLearner.navigateToLearnerDashboardAsLoggedInUser();
     await loggedInLearner.navigateToProgressSection();
 
     await loggedInLearner.expectScreenshotToMatch(
