+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/provide-feedback-on-the-lesson-or-report-it-from-the-lesson-player.spec.ts
@@ -112,7 +112,7 @@ test.describe('Logged-In Learner', function () {
     await loggedInLearner.navigateToCommunityLibraryPage();
     await loggedInLearner.searchForLessonInSearchBar('Algebra Basics');
     await loggedInLearner.playLessonFromSearchResults('Algebra Basics');
-    await loggedInLearner.continueToNextCard();
+    await loggedInLearner.continueToNextCardAsLoggedOutUser();
 
     // Report Exploration.
     await loggedInLearner.reportExploration('It is an ad');
