+++ core/tests/playwright-acceptance-tests/specs/logged-in-learner/reaches-a-checkpoint-and-saves-their-progress.spec.ts
@@ -112,9 +112,9 @@ test.describe('Logged-in User', function () {
     await loggedInUser.playLessonFromSearchResults('Positive Numbers');
 
     // Continue to the next card and submit an answer.
-    await loggedInUser.continueToNextCard();
+    await loggedInUser.continueToNextCardAsLoggedOutUser();
     await loggedInUser.submitAnswer('-25');
-    await loggedInUser.continueToNextCard();
+    await loggedInUser.continueToNextCardAsLoggedOutUser();
 
     // Verify that the checkpoint modal appears and reload the page.
     await loggedInUser.verifyCheckpointModalAppears();
@@ -126,7 +126,7 @@ test.describe('Logged-in User', function () {
     await loggedInUser.expectProgressReminder(true);
     await loggedInUser.chooseActionInProgressRemainder('Resume');
 
-    await loggedInUser.continueToNextCard();
+    await loggedInUser.continueToNextCardAsLoggedOutUser();
     await loggedInUser.expectCardContentToMatch(
       'Lesson completed successfully. We have practiced negative numbers.'
     );
@@ -140,9 +140,9 @@ test.describe('Logged-in User', function () {
     // Continue the exploration from where they left off.
     await loggedInUser.chooseActionInProgressRemainder('Restart');
 
-    await loggedInUser.continueToNextCard();
+    await loggedInUser.continueToNextCardAsLoggedOutUser();
     await loggedInUser.submitAnswer('-99');
-    await loggedInUser.continueToNextCard();
+    await loggedInUser.continueToNextCardAsLoggedOutUser();
   });
 
   test.afterAll(async function () {
