+++ core/tests/playwright-acceptance-tests/specs/logged-out-learner/listen-to-voiceovers-of-the-lessons-in-the-lesson-player.spec.ts
@@ -219,7 +219,7 @@ test.describe('Logged-Out Learner', function () {
     await loggedOutLearner.expectVoiceoverIsPlayable(false);
 
     // Check audio (voiceover) avaibility in next card.
-    await loggedOutLearner.continueToNextCard();
+    await loggedOutLearner.continueToNextCardAsLoggedOutUser();
     await loggedOutLearner.expectVoiceoverIsPlayable();
 
     // Play Voiceovers.
@@ -230,17 +230,17 @@ test.describe('Logged-Out Learner', function () {
 
   test('should be able to change the audio language', async function () {
     // Play voiceovers in Hindi.
-    await loggedOutLearner.playExploration(explorationId);
+    await loggedOutLearner.playExplorationAsLoggedOutUser(explorationId);
     await loggedOutLearner.changeLessonLanguage('hi');
 
-    await loggedOutLearner.continueToNextCard();
+    await loggedOutLearner.continueToNextCardAsLoggedOutUser();
     await loggedOutLearner.expectVoiceoverIsPlayable();
   });
 
   test('should be able to skip some parts of audio', async function () {
     await loggedOutLearner.reloadPage();
     await loggedOutLearner.changeLessonLanguage('hi');
-    await loggedOutLearner.continueToNextCard();
+    await loggedOutLearner.continueToNextCardAsLoggedOutUser();
 
     await loggedOutLearner.expectVoiceoverIsSkippable();
   });
