+++ core/tests/playwright-acceptance-tests/specs/logged-out-learner/share-the-lesson-from-the-lesson-player.spec.ts
@@ -116,8 +116,8 @@ test.describe('Logged-Out Learner', function () {
   });
 
   test('should be able to share the lesson using copy link', async function () {
-    await loggedOutUser.playExploration(explorationId);
-    await loggedOutUser.continueToNextCard();
+    await loggedOutUser.playExplorationAsLoggedOutUser(explorationId);
+    await loggedOutUser.continueToNextCardAsLoggedOutUser();
 
     await loggedOutUser.generateAttribution();
     await loggedOutUser.expectAttributionInHtmlSectionToBe(
