+++ core/tests/playwright-acceptance-tests/utilities/user/logged-in-user.ts
@@ -20,6 +20,7 @@ import {Page, expect, ElementHandle} from '@playwright/test';
 import {BaseUser} from '../common/playwright-utils';
 import testConstants from '../common/test-constants';
 import {showMessage} from '../common/show-message';
+import {NavigationUtils} from '../common/navigation-utils';
 
 const baseUrl = testConstants.URLs.BaseURL;
 const contributorDashboardAdminUrl =
@@ -27,12 +28,9 @@ const contributorDashboardAdminUrl =
 const learnerDashboardUrl = testConstants.URLs.LearnerDashboard;
 const profilePageUrlPrefix = testConstants.URLs.ProfilePagePrefix;
 const loginPageUrl = testConstants.URLs.Login;
-const moderatorPageUrl = testConstants.URLs.ModeratorPage;
-const releaseCoordinatorPageUrl = testConstants.URLs.ReleaseCoordinator;
 const signUpEmailField = testConstants.SignInDetails.inputField;
 const siteAdminPageUrl = testConstants.URLs.AdminPage;
 const splashPageUrl = testConstants.URLs.splash;
-const topicsAndSkillsDashboardUrl = testConstants.URLs.TopicAndSkillsDashboard;
 
 // Auth Pages selectors.
 const loginPage = '.e2e-test-login-page';
@@ -741,18 +739,9 @@ export class LoggedInUser extends BaseUser {
    * Verifies that the current page URL includes the expected page pathname.
    * @param {string} expectedPage - The expected page pathname (e.g., 'learner-dashboard').
    */
-  async expectToBeOnPage(expectedPage: string): Promise<void> {
-    await this.waitForStaticAssetsToLoad();
-    const url = this.page.url();
-
-    // Replace spaces in the expectedPage with hyphens.
-    const expectedPageInUrl = expectedPage.replace(/\s+/g, '-');
-
-    if (!url.toLowerCase().includes(expectedPageInUrl.toLowerCase())) {
-      throw new Error(
-        `Expected to be on page ${expectedPage}, but found ${url}`
-      );
-    }
+  async expectToBeOnPageAsLoggedInUser(expectedPage: string): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.expectToBeOnPage(expectedPage);
   }
 
   /**
@@ -893,11 +882,13 @@ export class LoggedInUser extends BaseUser {
 
   /**
    * Navigates to the learner dashboard.
+   * @param {boolean} verifyUrl - Whether to verify the URL after navigation. Defaults to true.
    */
-  async navigateToLearnerDashboard(): Promise<void> {
-    await this.goto(learnerDashboardUrl);
-    await this.waitForPageToFullyLoad();
-    await this.expectElementToBeAttachedInDOM(homeTabSectionInLearnerDashboard);
+  async navigateToLearnerDashboardAsLoggedInUser(
+    verifyUrl: boolean = true
+  ): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.navigateToLearnerDashboard(verifyUrl);
   }
 
   /**
@@ -1239,9 +1230,13 @@ export class LoggedInUser extends BaseUser {
 
   /**
    * Navigates to the Moderator page.
+   * @param {boolean} verifyUrl - Whether to verify the URL after navigation. Defaults to true.
    */
-  async navigateToModeratorPage(): Promise<void> {
-    await this.goto(moderatorPageUrl);
+  async navigateToModeratorPageAsLoggedInUser(
+    verifyUrl: boolean = true
+  ): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.navigateToModeratorPage(verifyUrl);
   }
 
   /**
@@ -1260,8 +1255,11 @@ export class LoggedInUser extends BaseUser {
   /**
    * Navigates to the Release Coordinator page.
    */
-  async navigateToReleaseCoordinatorPage(): Promise<void> {
-    await this.goto(releaseCoordinatorPageUrl);
+  async navigateToReleaseCoordinatorPageAsLoggedInUser(
+    verifyUrl: boolean = true
+  ): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.navigateToReleaseCoordinatorPage();
   }
 
   /**
@@ -1289,8 +1287,9 @@ export class LoggedInUser extends BaseUser {
   /**
    * Navigates to the Topics and Skills Dashboard page.
    */
-  async navigateToTopicsAndSkillsDashboardPage(): Promise<void> {
-    await this.goto(topicsAndSkillsDashboardUrl);
+  async navigateToTopicsAndSkillsDashboardPageAsLoggedInUser(): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.navigateToTopicsAndSkillsDashboardPage();
   }
 
   /**
@@ -1980,22 +1979,22 @@ export class LoggedInUser extends BaseUser {
    * Navigates to the splash page.
    * @param {string} expectedURL - The expected URL after navigation. Defaults to `${baseUrl}/`.
    */
-  async navigateToSplashPage(
+  async navigateToSplashPageAsLoggedInUser(
     expectedURL: string = learnerDashboardUrl
   ): Promise<void> {
-    // We explicitly check for expected URL instead of verifying it through
-    // BaseUser.goto as /splash redirects user to a different page.
-    await this.goto(splashPageUrl, false);
-
-    expect(this.page.url()).toBe(expectedURL);
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.navigateToSplashPage(expectedURL);
   }
 
   /**
    * Navigates to the exploration page and starts playing the exploration.
    * @param {string} explorationId - The ID of the exploration to play.
    */
-  async playExploration(explorationId: string | null): Promise<void> {
-    await this.goto(`${baseUrl}/explore/${explorationId as string}`);
+  async playExplorationAsLoggedInUser(
+    explorationId: string | null
+  ): Promise<void> {
+    const navigationUtils = new NavigationUtils(this);
+    await navigationUtils.playExploration(baseUrl, explorationId);
   }
 
   /**
